+++ SDL/PixelTriplet.cuh
@@ -47,7 +47,7 @@ namespace SDL
         bool* partOfPT5;
 
         //for track extensions
-        unsigned int* logicalLayers;
+        uint8_t* logicalLayers;
         unsigned int* hitIndices;
         uint16_t* lowerModuleIndices;
         FPX* centerX;
